Учебный проект: время разбивать камни (Разбить на модули)
