+++ src/components/templates/boardButtonTemplate.js
@@ -1,0 +1,5 @@
+const boardButtonTemplate = () => {
+  return `<button class="load-more" type="button">load more</button>`;
+};
+
+export default boardButtonTemplate();
