+++ src/components/templates/boardTasksList.js
@@ -1,0 +1,12 @@
+import boardTaskTemplate from './boardTaskTemplate.js'
+
+const boardTasksList = () => {
+  let taskCount = 3;
+  let boardTasks = '';
+  for (let i = 0; i < taskCount; i++) {
+    boardTasks += boardTaskTemplate;
+  }
+  return boardTasks;
+};
+
+export default boardTasksList();
